+++ CHANGELOG.md
@@ -8,6 +8,7 @@ X.Y.Z Release notes
 * Added Progress API `realm.syncSession.addProgressNotification` and `realm.syncSession.removeProgressNotification`
 * Added additional parameter for `Realm.open` and `Realm.openAsync` for download progress notifications 
 * Added `Realm.deleteFile` for deleting a Realm (#363).
+* Added `Realm.deleteModel` for deleting a Realm model in a migration (#573).
 
 ### Bug fixes
 * Adding missing TypeScript declation (#1283).